open_aoi/src/open_aoi_core/requirements.txt
opencv-python
python-dotenv
nicegui
sqlalchemy
pymysql
cryptography
bcrypt
pillow
minio